Fix variable collision for `EXISTS` join (#2047)
So far, an `EXISTS` join with a property path on each side fails because the hidden variable from the left side is reused on the right side. This is now fixed by hiding variables on the right side as if the `EXISTS` query was a subquery. Fixes #1813

NOTE: A cleaner solution would avoid colliding variable names in the first place. But that requires a larger refactoring, since the implementation of `ExistsJoin` currently creates a new `QueryPlanner` instance for its contents.