Consider `GRAPH` variable inside of `GRAPH` clause (#2135)
Queries like the following did not work correctly so far because the `GRAPH` variable was treated like any other variable inside of the `GRAPH` clause. This is now tentatively fixed by adding a suitable `FILTER` operation.
```sparql
SELECT * {
  GRAPH ?g {
    ?g ?p ?o
  }
}
```
NOTE: There are still problems when the `?g` in the example above is also used in other patterns inside the `GRAPH` clause, like `BIND` or `VALUES`.  To properly fix this, we need an operation that efficiently returns a list of all distinct graphs, which we do not have yet (but plan to have soon).