Quick fix for handling `LIMIT` and `OFFSET` in Cartesian product join (#2148)
In the current code, the `CartesianProductJoin` operation dynamically sets the limit and offset for its children and then resets them to their original values. However, this mechanism is flawed because the resetting happens too early for lazily evaluated children, which then falsely operate with the reset values. This is now tentatively fixed in two ways. First, the limit and offset of the children is no longer reset. Second, an error is thrown when the children are evaluated again (which is meant as a safeguard and cannot happen when processing real SPARQL queries). In particular, this fixes #2146 

NOTE: A proper fix requires a refactoring of the handling of limits and offsets. This is work for a separate PR.