Fix bug in parallel Turtle parser for small input files (#2159)
The parallel Turtle parser dynamically splits the input into batches and has to take care that predicate-object lists aka triple statements are not split in the middle. To that end, the parser searches for a particular regex, which is supposed to match the final `.` of a triple statement but can also match in the middle of the prefix declarations. In particular, this can happen when there are very many such declarations or when the input is very small. But so far, the parser only accepted prefix declarations in the first batch. This is now fixed (but it is still required that all prefix declaration come before the first triple statement).

In particular, fixes #2113

Co-authored by: Johannes Kalmbach <kalmbach@cs.uni-freiburg.de>