Use `ScanSpecAndBlocks` in `IndexScan` to avoid redundant copying of metadata (#2084)
This continues the work started in #1958 (see the description there) by now using `ScanSpecAndBlocks` throughout the whole call chain for performing an index scan and thus avoid redundant copies of the block metadata. In particular, this makes query planning faster for binary filters that have to read many blocks in the end. Specifically:

1. Each `IndexScan` now contains a `ScanSpecAndBlocks`, which in turn contains `BlockMetadataRanges` = a sequence of subranges of the full `CompressedBlockMetadata`
2. The `PrefilterExpression`, if available, is now evaluated only once when passed to the `IndexScan`, updating the `BlockMetadataRanges`
3. These `BlockMetadataRanges` are then used for computing the result *and* the size estimate of the `IndexScan`