Refactor `MINUS` implementation and fix handling of `UNDEF` (#2131)
This changes the implementation of the `MINUS` operation to use the interface provided by `JoinAlgorithms.h` and such that `UNDEF` values are handled correctly. Here are two example queries, which gave a wrong result so far and are now fixed. Note that according to the SPARQL 1.1 standard, `MINUS` eliminates a solution mapping for the LHS only if there is a solution mapping for the RHS such that all shared bindings match *and there is at least one shared bound variable*.
```sparql
SELECT * WHERE {
  VALUES (?a ?b) { (UNDEF UNDEF) (UNDEF 1) }
  MINUS {
    VALUES (?a ?b) { (1 UNDEF) (UNDEF UNDEF) (1 1) }
  }
  # Should result in a single with UNDEF UNDEF
}
```
```sparql
SELECT * WHERE {
  VALUES (?a ?b) { (UNDEF UNDEF) (UNDEF 1) }
  MINUS {
    VALUES (?a ?b) { (1 UNDEF) (UNDEF UNDEF) }
  }
  # Should result in two rows: one with UNDEF UNDEF, and one with UNDEF 1
}
```