Have a single counter per SPARQL query for internal variables (#2048)
The following query led to an assertion failure so far, for the following reason. Line 2 and 3 both create an internal variable (due to the `[]` in this case, but internal variables can also be created for other reasons, e.g., property paths). The generic name for this variable is created using an internal counter. So far, a separate internal counter was used for subqueries and `EXIST`. This is now fixed by using a single internal counter for the whole SPARQL query. In particular, fixes #2034.
```
SELECT * WHERE {
  ?a a [] .
  { SELECT * WHERE {?a a []}}
}
```
NOTE: This change increase the code complexity somewhat, but we found this preferable to trying to coordinate multiple internal counters for a single SPARQL query.