Optimize export performance by reusing a fixed-size buffer (#2155)
A query result is internally a table of fixed-size `Id`s and has to be converted to a string representation when exporting it with one of the standard media types. So far, QLever did this using `std::ostringstream`, which requires many (typically small) memory allocations. This is now improved by reusing a fixed-size buffer.

NOTE: When exporting 10 M triples from Wikidata, this makes the export around 10% faster. This is just a first step and there is a lot of room for further improvement. For example, the relevant methods currently returns `cppcoro::generator<std::string>`, which is an obstacle for even more reuse and thus saving even more allocations.