Add type trait `anyIsInstantiationOf` (#2222)
Add the type trait `anyIsInstantiationOf<T, Ts...>`, which checks if any type in `Ts...` is an instantiation of the type `T`. For example, the following holds because  `std::vector<int>` in `Ts...` is an instantiation of `std::vector`.
```cpp
anyIsInstantiationOf<std::vector, std::tuple<bool, int>, std::vector<int>, std::pair<bool, bool>>
```
Also clean up comments and provide examples in the code in the form of `static_assert` statements. This is preparation for #2221 and #2205.