Prefilter `SpatialJoin` input using bounding boxes from `GeoVocabulary` (#2205)
For a `SpatialJoin` operation, use the precomputed bounding boxes (available if the index was built with vocabulary type `on-disk-compressed-geo-split`) to reduce the number of rows of the larger operand before the actual spatial join is computed. Specifically, eliminate all rows from the larger operand where the bounding box does not intersect the aggregate bounding box of all geometries from the smaller operand. Builds on #2221, #2222, and #2223

For example, for the following query (all objects in the city of Zürich) on the OSM data of Switzerland, this approximately halves the time for the spatial join, from `8 s` to `4 s`. The reason that the improvement is not more dramatic is that this optimization saves on reading many full geometries from disk, but the bounding boxes also have to be read from disk and the average number of points of a full geometry is only around 10.
```
SELECT ?any ?any_geom WHERE {
  osmrel:1682248 geo:hasGeometry/geo:asWKT ?zurich_geom .
  ?any geo:hasGeometry/geo:asWKT ?any_geom .
  FILTER geof:sfContains(?zurich_geom,?any_geom)
}
```