Implement index nested loop join for `MINUS` and `EXISTS` (#2251)
Queries like the following two run out of memory with the current code because the RHS (the argument of the `MINUS` or `EXISTS`) is too large to be materialized and sorted in memory (and QLever currently does *not* have a query operation that sorts lazily with limited memory). This is now remedied by an index nested loop join, provided that the size estimate of the LHS is smaller than (or equal to) the size estimate of the RHS. Specifically, the LHS is then fully materialized (hence the precondition on its size) and binary-searched for each element from the lazily produced RHS.

NOTE: This has running time ~ RHS * log(LHS), which is large but at least it works now. A better algorithm could be to sort chunks of the RHS and join them with the LHS. However, that is not trivial for multi-column joins and when a single element matches mutliple blocks.
```sparql
SELECT * WHERE {
  ?s wdt:P31 ?o .
  MINUS {
    ?s wdt:P31 ?m .
    ?m wdt:P279* ?o .
  }
}
```
```sparql
SELECT * WHERE {
  ?s wdt:P31 ?o .
  FILTER EXISTS {
    ?s wdt:P31 ?m .
    ?m wdt:P279* ?o .
  }
}
```