"Multiply out" all `Union` operations + improve the cost estimate (#2267)
So far, a "multiplied out" version of the `Union` (that is, where the context is "pulled" into both the LHS and RHS of the operation) was added as a possible query plan only when there was a transitive path in either child. Now, the multiplied versions are always added. At the same time, improve the cost estimate by taking into account whether the result has to be sorted or not. If not, the `Union` is cheap because the results of the LHS and RHS merely have to be concatenated. If the result has to be sorted, the `Union` is expensive.