Implement index nested loop join for `OPTIONAL` (#2265)
Follow-up to #2251, which implements index nested loop joins for `OPTIONAL`, under the same conditions as in #2251.  Here is an analogous example query, which fails with an out-of-memory error with the current code but now completes, albeit slowly. On the side, fix a minor bug in the algorithm for computing a lazy optional join, which occurs when an undefined value on the left is joined with an empty table on the right.
```sparql
SELECT (COUNT(*) AS ?count) {
  ?a wdt:P2860 ?b .
  OPTIONAL {
    ?b wdt:P31 ?c .
    ?c schema:description ?d .
  }
}
```