Use consistent `ANTLR` syntax for character ranges (#2230)
Since #2215, two different syntaxes are used for character ranges (the focus of that PR was on something else). For example, for capital letters there is `'A'..'Z'` and `[A-Z]`. Only the latter supports Unicode codepoints greater than `U+FFFF`, so we now use that throughout. On the side, bring the whole grammar visually closer to the grammar provided by the SPARQL standard.