First version of `libqlever` for an embedded use of QLever (#2100)
This enables writing C++ code that builds an index for a given dataset, loads such an index from disk, and issues queries in an embedded fashion, that is, without running a HTTP server. For a usage example, have a look at `src/libqlever/LibQLeverExample.cpp`. This can be compiled with `ninja -C build LibQLeverExample`, which produces a binary `LibQLeverExample` in the `build` folder. To compile other code in this vein, just add the corresponding lines to `CMakeLists.txt` in the top-level folder (search for `LibQLeverExample`). Some notes and limitations:

1. This is a first minimal but functional and useful version. For an overview of what can be configured already now, see `src/libqlever/Qlever.h`
2. To get detailed timings on the various parts of the query execution (like when you click "Analysis" in the QLever UI), use `MediaType::qleverJson`
3. To compile code like that in `LibQLeverExample.cpp`, QLever's HTTP server code needs to be neither compiled nor linked
4. The code for `IndexBuilderMain` already uses `libqlever` now and has become much simpler
5. Some of the current limitations: timeouts, manual cancellation, live timings while the query is running, updates. These are all supported by QLever in the standard client/server setup, but not yet by `libqlever` 

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>