Add CMake option `REDUCED_FEATURE_SET_FOR_CPP17` (#2387)
The option is `OFF` by default, in which case the same code (with the same functionality) as before is compiled. When the option is set to `ON`, then certain parts of the code are excluded from compilation or linking and the following functionality is **not** available:

1. The operations `Service` and `Load` (which require the `http` library, which is not linked)
2. Transitive property paths
3. Building a full-text index
4. The binaries `ServerMain` and `IndexBuilderMain` (but `libqlever` **is** available)

The point of the option is to reduce the amount of code that has to be migrated to C++17, yet preserving core functionality that is sufficient for many applications.

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>