Migrate `std::source_location` to C++17 (#2406)
Implement `ad_utility::source_location` , a drop-in replacement for `std::source_location`.  The function `source_location::current()` is replaced by a macro `AD_CURRENT_SOURCE_LOC()` which has the exact same behavior. In particular, when used as a defaulted argument to a function, it provides the source location from which the function was called.