Migrate default-constructed lambdas to C++17 (#2417)
In C++20, capture-less lambdas of the form `[](someArgs){ doSomething; }` can be default-constructed, in C++17 they can't. This is relevant when a class template takes the type of a callable (using `decltype`), and then default-constructs it.  Several of these lambdas are now rewritten, each to an equivalent named `struct` with an `operator()`.