Update SERVER_URL in workflow configuration (#2432)
The workflow that uploads the SPARQL conformance results was broken by the recent move from `qlever.cs.uni-freiburg.de` to `qlever.dev`, because it did not honor HTTP redirect repsonse codes. This PR makes the uploader work in the presence of redirects, and also updates the URL of the upload server.