Add `geof:length` and `geof:metricLength` using `GeometryInfo` (#2172)
Add support for the GeoSPARQL functions `geof:length` and `geof:metricLength`. For a `LINESTRING` or a `POLYGON` this is the sum of the length of all its segments. For more complex shapes it is the sum of the lengths of the simpler shapes forming the complex shape. For a `POINT` it is zero. For example (length of the Rhine Valley railway between Mannheim and Basel):
```sparql
SELECT * {
  osmrel:1781296 geo:hasGeometry/geo:asWKT ?geom .
  BIND (geof:metricLength(?geom) AS ?meters) # Result: 267084
  BIND (geof:length(?geom, unit:KiloM) AS ?kilometers) # Result: 267.084
  BIND (geof:length(?geom, unit:MI) AS ?miles) # Result: 165.958
}
```
On the side, enforce invariants on the helper structs for `GeometryInfo`.

NOTE: This is an index-breaking change with regard to the `on-disk-compressed-geo-split` vocabulary type. It will be merged in a sequence with #2238 and #2430